refactor(cli): move run command top level, debug mode sub chain
