+++ cli/src/cli/commands/node/mod.rs
@@ -8,7 +8,6 @@ mod get_nodes;
 mod node_result;
 mod register_node;
 pub use node_result::NodeResult;
-mod run;
 mod unregister_node;
 mod update_node;
 
@@ -17,20 +16,19 @@ pub enum Node {
     GetNode(get_node::GetNode),
     GetNodes(get_nodes::GetNodes),
     RegisterNode(register_node::RegisterNode),
-    Run(run::Run),
     UpdateNode(update_node::UpdateNode),
     UnregisterNode(unregister_node::UnregisterNode),
 }
 
 impl Node {
-    pub fn handle(self, config: AppConfig) -> Result<()> {
+    #[tokio::main]
+    pub async fn handle(self, config: AppConfig) -> Result<()> {
         match self {
-            Self::GetNode(get_node) => get_node.handle(),
-            Self::GetNodes(get_nodes) => get_nodes.handle(),
-            Self::RegisterNode(register_node) => register_node.handle(),
-            Self::Run(run) => run.handle(config),
-            Self::UpdateNode(update_node) => update_node.handle(),
-            Self::UnregisterNode(unregister_node) => unregister_node.handle(),
+            Self::GetNode(get_node) => get_node.handle().await,
+            Self::GetNodes(get_nodes) => get_nodes.handle().await,
+            Self::RegisterNode(register_node) => register_node.handle().await,
+            Self::UpdateNode(update_node) => update_node.handle().await,
+            Self::UnregisterNode(unregister_node) => unregister_node.handle().await,
         }
     }
 }
