+++ cli/src/cli/commands/node/node_result.rs
@@ -1,5 +1,6 @@
 use serde::{Deserialize, Serialize};
 
+// TODO move to sdk cause this can also be used in the WASM
 #[derive(Debug, Clone, Eq, PartialEq, Deserialize, Serialize)]
 pub struct NodeResult {
     pub owner:          String,
