+++ cli/src/cli/commands/node/update_node.rs
@@ -28,7 +28,6 @@ pub struct UpdateNode {
 }
 
 impl UpdateNode {
-    #[tokio::main]
     pub async fn handle(self) -> Result<()> {
         todo!("chain view call");
         return Ok(());
