+++ cli/src/cli/commands/sub_chain/mod.rs
@@ -8,11 +8,12 @@ mod view;
 
 #[derive(Debug, Subcommand)]
 pub enum SubChain {
-    // ./seda sub-chain view near mc.mennat0.testnet get_nodes "{\"offset\":\"0\",\"limit\":\"2\"}"
-    View(view::View),
     // ./seda sub-chain call near mc.mennat0.testnet register_node
     // "{\"socket_address\":\"127.0.0.1:8080\"}" "870000000000000000000"
     Call(call::Call),
+    // TODO maybe a ListMethods,
+    // ./seda sub-chain view near mc.mennat0.testnet get_nodes "{\"offset\":\"0\",\"limit\":\"2\"}"
+    View(view::View),
 }
 
 impl SubChain {
